Fix `test_runtime_env_container` (#54330)
Move `httpx` out of `test_utils` because for some reason it is not
available in the image used for `test_runtime_env_container.
